Some small convenience changes

+++ update_and_clean_archlinux.sh
@@ -7,8 +7,10 @@
 # creates a log of update completion times and start times
 #
 
-LOG_FILE=${PWD}/update.log # log file location
+this_dir=$(dirname "$0") #this files directory
+LOG_FILE=${this_dir}/update.log # log file location
 DATE=$(date) # current date
+DIV="-------------------------------------------"
 
 echo "Starting update"
 
@@ -17,17 +19,23 @@ echo "Starting update"
 if [ -f "LOG_FILE" ]; then
   echo "Log file does not exist, creating..."
   touch ${LOG_FILE}
-  echo "Update started at: " $DATE >> $LOG_FILE
+  echo "Update started at: " ${DATE} >> ${LOG_FILE}
+  echo "Update started at: " ${DATE} 
+  echo -e "${DIV}\n"
   sudo pacman -Syu --noconfirm 
   trizen -Syu --noconfirm --noedit
   sudo pacman -Sc #clear cache and unnecessary packages
   sudo pacman -R $(pacman -Qqtd) #clean orphaned packages
-  echo "Update finished at: " $DATE >> $LOG_FILE
+  echo "${DIV}"
+  echo "Update finished at: " ${DATE} >> ${LOG_FILE}
 else
-  echo -e "Update started at: " $DATE >> $LOG_FILE
+  echo -e "Update started at: " ${DATE} >> ${LOG_FILE}
+  echo -e "${DIV}\n"
   sudo pacman -Syu --noconfirm 
   trizen -Syu --noconfirm --noedit
-  echo "Y" | sudo pacman -Sc #clear cache and unnecessary packages
-  sudo pacman -R $(pacman -Qqtd) #clean orphaned packages
-  echo "Update finished at: " $DATE >> $LOG_FILE
+  yes | sudo pacman -Sc #clear cache and unnecessary packages
+  yes | sudo pacman -R $(pacman -Qqtd) #clean orphaned packages if needed
+  echo ${DIV}
+  echo "Update finished at: " ${DATE} >> ${LOG_FILE}
+  echo "Update finished at: " ${DATE} 
 fi
